requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorlog==6.4.1
hikari==2.0.0.dev102
hikari-lightbulb==1.4.0
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pure25519==0.0.1
python-dotenv==0.19.0
regex==2021.9.30
requests==2.26.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
